CM-34503 - Add Secrets Violation Card
