+++ src/panels/violation/card/sca.ts
@@ -1,0 +1,45 @@
+export default `
+<section class="card">
+  <section class="header">
+      <div class="severity">
+          <img class="severity-icon" alt="Severity" />
+      </div>
+      <div class="title">None</div>
+      <div class="short-details"></div>
+  </section>
+
+  <section class="hr details">
+      <div class="details-item">
+        <div class="details-item-title">Package:</div>
+        <div class="details-item-value package">None</div>
+      </div>
+      <div class="details-item">
+        <div class="details-item-title">Version:</div>
+        <div class="details-item-value version">None</div>
+      </div>
+      <div class="first-patched-version">
+        <div class="details-item">
+          <div class="details-item-title">First patched version:</div>
+          <div class="details-item-value first-patched-version-value">None</div>
+        </div>
+      </div>
+      <div class="dependency-path">
+        <div class="details-item">
+          <div class="details-item-title">Dependency path:</div>
+          <div class="details-item-value dependency-path-value">None</div>
+        </div>
+      </div>
+      <div class="licence">
+        <div class="details-item">
+          <div class="details-item-title">License:</div>
+          <div class="details-item-value licence-value">None</div>
+        </div>
+      </div>
+  </section>
+
+  <section class="hr summary">
+      <h2>Summary</h2>
+      <div class="summary-text">None</div>
+  </section>
+</section>
+`;
