+++ src/panels/violation/card/secret.ts
@@ -1,0 +1,36 @@
+export default `
+<section class="card">
+  <section class="header">
+      <div class="severity">
+          <img class="severity-icon" alt="Severity" />
+      </div>
+      <div class="title">None</div>
+      <div class="short-details"></div>
+  </section>
+
+  <section class="hr details">
+      <div class="details-item">
+        <div class="details-item-title">Rule ID:</div>
+        <div class="details-item-value rule">None</div>
+      </div>
+      <div class="details-item">
+        <div class="details-item-title">In file:</div>
+        <div class="details-item-value file">None</div>
+      </div>
+      <div class="details-item">
+        <div class="details-item-title">Secret SHA:</div>
+        <div class="details-item-value sha">None</div>
+      </div>
+  </section>
+  
+  <section class="hr">
+    <h2>Summary</h2>
+    <div class="summary-text">None</div>
+  </section>
+
+  <section class="company-guidelines">
+      <h2>Company Guidelines</h2>
+      <div class="company-guidelines-text">None</div>
+  </section>
+</section>
+`;
